requirements.txt
PyQt6
spotipy
google-generativeai
python-dotenv